Tower Music
Background music for the tower. It is able to keep playing between tower1 and tower2. However there is a minor hiccup where if you go back down to tower1 it will start a new track and two of the bgm will be playing. Will work on fixing that when I can